consumer/requirements.txt
anyio==4.7.0
attrs==24.3.0
avro-python3==1.10.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
confluent-kafka==2.7.0
elastic-transport==8.15.1
elasticsearch==8.17.0
fastavro==1.10.0
finnhub-python==2.4.22
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kafka-python==2.0.2
numpy==2.2.0
pandas==2.2.3
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.2
